+++ packages/jwt/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  collectCoverageFrom: ["src/**/*.{js,jsx,ts,tsx}"],
+  roots: ["<rootDir>/src/"],
+  clearMocks: true,
+  resetMocks: true,
+  testEnvironment: "node",
+  moduleNameMapper: {
+    "@/(.*)": "<rootDir>/src/$1",
+  },
+};