+++ packages/jwt/src/JWTUser.ts
@@ -1,4 +1,9 @@
-import { IJWTUser, JWTPayload } from "./types";
+import { BitmaskUtils } from "./BitmaskUtils";
+import { IPermissionService } from "./PermissionService";
+import { PermissionDefinitionNotFound } from "./errors/PermissionNotFound";
+import { PermissionServiceNotInstantiated } from "./errors/PermissionServiceNotInstantiated";
+import { IJWTUser, JWTPayload, Permission } from "./types";
+import { NoPermissionError } from "./errors/NoPermissionError";
 
 export class JWTUser implements IJWTUser {
   id: string;
@@ -14,6 +19,30 @@ export class JWTUser implements IJWTUser {
   permissions: string;
 
   segmentReference: string | null;
+
+  /**
+   * Static reference to the permission service
+   * This is used to make sure that the permission service is only instantiated once
+   * and can be used by all instances of JWTUser
+   */
+  static #permissionService: IPermissionService | null = null;
+
+  /**
+   * Instantiate the permission service for all instances of JWTUser
+   */
+  static setPermissionService(service: IPermissionService): void {
+    JWTUser.#permissionService = service;
+  }
+
+  /**
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  static getPermissionService(): IPermissionService {
+    if (!JWTUser.#permissionService) {
+      throw new PermissionServiceNotInstantiated();
+    }
+    return JWTUser.#permissionService;
+  }
 
   constructor({
     id,
@@ -75,11 +104,117 @@ export class JWTUser implements IJWTUser {
 
   /**
    * Check if user has permission to perform an action
-   * @note - not implemented, will be implemented in the next release
    * @param permissionCode - permission code to check
    * @returns boolean
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
    */
-  public hasPermission(_permissionCode: string): boolean {
-    throw new Error("Not implemented");
+  public async hasPermission(permissionCode: string): Promise<boolean> {
+    const definition = await this.getPermissionDefinition(permissionCode);
+    const checkBitmask = BitmaskUtils.encodeFromPosition(definition.position);
+    const permissionsBitmask = BitmaskUtils.decodeFromBase64(this.permissions);
+    return BitmaskUtils.checkBitmask(permissionsBitmask, checkBitmask);
+  }
+
+  /**
+   * Throws an error if the user does not have the required permission
+   * @param permissionCode
+   * @throws {NoPermissionError}
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  public async validatePermission(permissionCode: string): Promise<void> {
+    const hasPermission = await this.hasPermission(permissionCode);
+    if (!hasPermission) {
+      throw new NoPermissionError([permissionCode]);
+    }
+  }
+
+  /**
+   * Check if user has all permissions to perform an action
+   * @param permissionCodes - permission codes to check
+   * @returns boolean
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  public async hasEveryPermission(permissionCodes: string[]): Promise<boolean> {
+    const permissionsNotMet: string[] = [];
+    for (const permissionCode of permissionCodes) {
+      const hasPermission = await this.hasPermission(permissionCode);
+      if (!hasPermission) {
+        permissionsNotMet.push(permissionCode);
+      }
+    }
+    return permissionsNotMet.length === 0;
+  }
+
+  /**
+   * Throws an error if the user does not have all permissions
+   * @param permissionCode
+   * @throws {NoPermissionError}
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  public async validateEveryPermission(
+    permissionCodes: string[]
+  ): Promise<void> {
+    const permissionsNotMet: string[] = [];
+    for (const permissionCode of permissionCodes) {
+      const hasPermission = await this.hasPermission(permissionCode);
+      if (!hasPermission) {
+        permissionsNotMet.push(permissionCode);
+      }
+    }
+    if (permissionsNotMet.length > 0) {
+      throw new NoPermissionError(permissionsNotMet);
+    }
+  }
+
+  /**
+   * Check if user has at least one permission
+   * @param permissionCodes - permission codes to check
+   * @returns boolean
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  public async hasSomePermission(permissionCodes: string[]): Promise<boolean> {
+    for (const permissionCode of permissionCodes) {
+      const hasPermission = await this.hasPermission(permissionCode);
+      if (hasPermission) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Throws an error if the user does not have at least one of the permissions
+   * @param permissionCode
+   * @throws {NoPermissionError}
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  public async validateSomePermission(
+    permissionCodes: string[]
+  ): Promise<void> {
+    const hasSomePermission = await this.hasSomePermission(permissionCodes);
+    if (!hasSomePermission) {
+      throw new NoPermissionError(permissionCodes);
+    }
+  }
+
+  /**
+   * @throws {PermissionDefinitionNotFound}
+   * @throws {PermissionServiceNotInstantiated}
+   */
+  private async getPermissionDefinition(
+    permissionCode: string
+  ): Promise<Permission> {
+    const definitions = await JWTUser.getPermissionService().getPermissions();
+    const definition = definitions.find((def) => def.code === permissionCode);
+    if (!definition) {
+      throw new PermissionDefinitionNotFound(permissionCode);
+    }
+    return definition;
   }
 }